first rough working oidc server
only optimal scenario is handled.

errors are not handled correctly, a lot of stuff remains to be done.

APIs are not clean, the global approach needs to be rethough
(is configurer/configuration/builder the way to go, do
we want to tie the configurer to Spring Security, ...).

Logout, optional bride to SAML RP, local STS are to be made, etc...