SpringViewProvider error handling
Temporary fix for SpringViewProvider:
- thymeleaf template calls response.getWriter().
- when a viewProvider generates an error, CXF
  calls response.getOutputStream()

This means that when there's an error during rendering
a thymeleaf template, servletEngine throws an error
(because both getWriter and getOutputStream were called).

With this temporary fix, SpringViewProvider dispatches
to /error and never throws an error.

This is temporary, the real fix should perhaps be done in CXF.